Further refactored PlotEffect class and initialised test implementation
If you're failing to understand PlotEffect: it exists as a stack of
modifier arrays where the array on the top is ALWAYS the set of
modifiers imposed by the effect itself. All other arrays below it in the
stack contain the original production modifiers held by affected tiles,
allowing for effects to be reverted later on by calling the revert()
function. Effects can be imposed either by calling
[PlotEffect.impose(LandPlot, mode)] directly, or by setting up its
internal Runnable object to call that after executing other processes
and running the Runnable instead.

Note that the productionModifiers array in the LandPlot class, which was
previously of type int[], was changed to be of type float[] so that it
could be better integrated into the new LandPlot class.